requirements.txt
environs==9.5.0
python-telegram-bot==13.13
requests==2.28.1